backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
pydantic>=2.5.0
python-multipart>=0.0.6
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.26.0
sentence-transformers>=2.2.2
scikit-surprise>=1.1.3
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
sqlalchemy>=2.0.23 